moved keys to a new file concerning objects
